Change type size in generate

+++ generator.c
@@ -11,23 +11,23 @@ int generate(int fd, EL_TYPE *buffer, size_t size,  uint64_t num_create){
 
 
 
-	int blocks = num_create/size;
+	uint64_t blocks = num_create/size;
 	if(num_create%size != 0){
 		blocks++;
 	}
 
-	int limit, offset;
+	uint64_t limit, offset;
 	offset = 0;
 
 	srand(time(NULL));
-	for(int i=0; i<blocks; i++){
+	for(uint64_t i=0; i<blocks; i++){
 		if(num_create>size){
 			limit = size;
 		}else{
 			limit = num_create;
 		}
 
-		for(int k=0; k<limit; k++){
+		for(uint64_t k=0; k<limit; k++){
 			buffer[offset+k] = rand();
 		}
